Refactor bezierbuilder to avoid `eval` calls.
This will help with pep8 naming convention compliance without breaking
the user-facing API. Specifically, we can lowercase the names of the
Bezier and CatmulClark functions without having to lowercase the string
argument values to the CLI.
